Inline heart element
