+++ src/main/java/by/jackraidenph/dragonsurvival/handlers/ItemsInit.java
@@ -3,7 +3,6 @@ package by.jackraidenph.dragonsurvival.handlers;
 import by.jackraidenph.dragonsurvival.DragonSurvivalMod;
 import by.jackraidenph.dragonsurvival.capability.DragonStateHandler;
 import by.jackraidenph.dragonsurvival.capability.DragonStateProvider;
-import by.jackraidenph.dragonsurvival.items.HeartElement;
 import by.jackraidenph.dragonsurvival.network.SyncLevel;
 import by.jackraidenph.dragonsurvival.util.DragonLevel;
 import net.minecraft.entity.SharedMonsterAttributes;
@@ -22,13 +21,43 @@ import net.minecraftforge.fml.network.PacketDistributor;
 
 @Mod.EventBusSubscriber(modid = DragonSurvivalMod.MODID, bus = Mod.EventBusSubscriber.Bus.MOD)
 public class ItemsInit {
-    public static HeartElement heartElement;
+    public static Item heartElement;
     public static Item starBone, elderDragonBone;
 
     @SubscribeEvent
     public static void register(final RegistryEvent.Register<Item> event) {
-        heartElement = new HeartElement(new Item.Properties().group(BlockInit.blocks).maxStackSize(16));
+        heartElement = new Item(new Item.Properties().group(BlockInit.blocks).maxStackSize(64)) {
+            @Override
+            public ActionResult<ItemStack> onItemRightClick(World worldIn, PlayerEntity playerIn, Hand handIn) {
+                LazyOptional<DragonStateHandler> dragonStateHandlerLazyOptional = playerIn.getCapability(DragonStateProvider.PLAYER_STATE_HANDLER_CAPABILITY);
+                if (dragonStateHandlerLazyOptional.isPresent()) {
+                    DragonStateHandler dragonStateHandler = dragonStateHandlerLazyOptional.orElseGet(() -> null);
+                    if (dragonStateHandler.isDragon()) {
+                        float maxHealth = playerIn.getMaxHealth();
+                        if (maxHealth < 40) {
+                            IAttributeInstance currentHealth = playerIn.getAttribute(SharedMonsterAttributes.MAX_HEALTH);
+                            currentHealth.setBaseValue(currentHealth.getBaseValue() + 2);
+                            playerIn.getHeldItem(handIn).shrink(1);
+
+                            if (currentHealth.getBaseValue() >= DragonLevel.ADULT.initialHealth) {
+                                dragonStateHandler.setLevel(DragonLevel.ADULT);
+                                if (!worldIn.isRemote)
+                                    DragonSurvivalMod.CHANNEL.send(PacketDistributor.ALL.noArg(), new SyncLevel(playerIn.getEntityId(), DragonLevel.ADULT));
+                            } else if (currentHealth.getBaseValue() >= DragonLevel.YOUNG.initialHealth) {
+                                dragonStateHandler.setLevel(DragonLevel.YOUNG);
+                                if (!worldIn.isRemote)
+                                    DragonSurvivalMod.CHANNEL.send(PacketDistributor.ALL.noArg(), new SyncLevel(playerIn.getEntityId(), DragonLevel.YOUNG));
+                            }
+
+                            return ActionResult.resultSuccess(playerIn.getHeldItem(handIn));
+                        }
+                    }
+                }
+                return super.onItemRightClick(worldIn, playerIn, handIn);
+            }
+        };
         heartElement.setRegistryName(DragonSurvivalMod.MODID, "heart_element");
+
         starBone = new Item(new Item.Properties().group(BlockInit.blocks)) {
             @Override
             public ActionResult<ItemStack> onItemRightClick(World worldIn, PlayerEntity playerIn, Hand handIn) {
@@ -57,6 +86,7 @@ public class ItemsInit {
                 return super.onItemRightClick(worldIn, playerIn, handIn);
             }
         }.setRegistryName(DragonSurvivalMod.MODID, "star_bone");
+
         elderDragonBone = new Item(new Item.Properties().group(BlockInit.blocks)).setRegistryName(DragonSurvivalMod.MODID, "elder_dragon_bone");
         event.getRegistry().registerAll(heartElement, starBone, elderDragonBone);
     }
